# Conflicts:
#	gen src/antlr/PivotParser.java
#	pom.xml

+++ src/main/java/compiler/Compiler.java
@@ -74,19 +74,25 @@ public class Compiler {
     private static void deleteOldGeneratedFiles(){
         File generatedFiles = new File(GENERATED_FILES_DIR);
 
-        // Remove all generated files
-        for (File f: Objects.requireNonNull(generatedFiles.listFiles())) {
-            if(f.isDirectory()){
-                try {
-                    FileUtils.deleteDirectory(f);
-                } catch (IOException e) {
-                    e.printStackTrace();
+
+        if(generatedFiles.exists()){
+            // Remove all generated files
+            for (File f: Objects.requireNonNull(generatedFiles.listFiles())) {
+                if(!f.getName().equals("TempToDelete") ){
+                    if(f.isDirectory()){
+                        try {
+                            FileUtils.deleteDirectory(f);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                    if(f.isFile()){
+                        f.delete();
+                    }
                 }
             }
-            if(f.isFile()){
-                f.delete();
-            }
         }
+
     }
 
     private static void generateJavaCode(Node ast){
